Moved tests into test dir and updated them to use chai
This should allow us more flexibility with expansion of the test
codebase.
